requirements.txt
aiohttp==3.7.2
asgiref==3.2.10
astroid==2.3.3
async-timeout==3.0.1
attrs==20.2.0
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
Click==7.0
colorama==0.4.3
cryptography==3.1
defusedxml==0.6.0
distro==1.5.0
dj-database-url==0.5.0
Django==3.1
django-allauth==0.42.0
django-cors-headers==3.4.0
django-filter==2.3.0
django-rest-auth==0.9.5
django-rest-knox==4.1.0
djangorestframework==3.11.1
geoip2==4.1.0
getmac==0.8.2
google-api-python-client==1.6.2
gunicorn==20.0.4
httpie==2.2.0
httplib2==0.18.1
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
lxml==4.6.1
maxminddb==2.0.3
mccabe==0.6.1
multidict==5.0.0
numpy==1.19.2
oauth2client==4.1.3
oauthlib==3.1.0
pandas==1.1.3
peewee==3.13.3
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.6.1
pylint==2.4.4
PyMySQL==0.9.3
pypykatz @ https://github.com/AlessandroZ/pypykatz/archive/master.zip
python-dateutil==2.8.1
python-decouple==3.3
python-nmap==0.6.1
python-shell==1.0.0
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
sentry-sdk==0.14.3
six==1.14.0
spotify-dl==7.0.0
spotipy==2.16.0
sqlparse==0.3.1
tabula-py==2.2.0
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.25.10
whitenoise==5.2.0
who-is-on-my-wifi==1.2.0
wrapt==1.11.2
yarl==1.6.2
youtube-dl==2020.11.1.1
